Extract out non-esbuild logic from esbuild utils
