+++ src/bundler/utils.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 
-import esbuild from "esbuild";
+import fs from "fs";
 
 export interface BundleOptions {
   minify: boolean;
@@ -22,19 +22,8 @@ export namespace BundleOptions {
   }
 }
 
-export function runBundler(
-  buildOptions: esbuild.BuildOptions,
-  envOptions?: BundleOptions
-) {
-  if (envOptions?.watch) {
-    runWatch(buildOptions).catch(console.warn);
-  } else {
-    esbuild.build(buildOptions).catch(console.warn);
-  }
-}
-
-async function runWatch(buildOptions: esbuild.BuildOptions) {
-  const ctx = await esbuild.context(buildOptions);
-  await ctx.watch();
-  console.log("Started esbuild watch!");
+export function getHash(): string {
+  const hash = fs.readFileSync("build/morcusnet.commit.txt").toString();
+  console.log(`Client commit hash: "${hash}"`);
+  return hash;
 }
